Split the CAPI into two crates to allow individualized panic handler impls
