+++ .gitignore
@@ -2,3 +2,4 @@
 /target/
 .idea
 ./target
+Cargo.lock
