+++ ekotrace-capi-impl/.gitignore
@@ -1,3 +1,4 @@
 target
-.idea
+*.json
 Cargo.lock
+.idea
