+++ ekotrace-capi-impl/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "ekotrace-capi-impl"
+version = "0.1.0"
+authors = ["Zachary Pierce <zack@auxon.io>"]
+edition = "2018"
+
+[lib]
+name = "ekotrace_capi_impl"
+# Note that we intentionally separate out the staticlib and cydlib
+# crate types into ekotrace-capi to avoid the pain of panic handler
+# definition conflicts
+crate-type = ["lib"]
+
+[dependencies]
+ekotrace = { path = "../ekotrace" }
+
+[build-dependencies]
